redeesign auth pages
